Completely Working
This repo is now completely working with signing into the app. Now going to add the profile and other useful features. I am now going to continue to create the other app that I am using this for. Going to add more to the UI/UX to make it more appealing. Going to add more objC support. Compinling in under 11 seconds trying to get it to under 10. Going to add more support in the future for more signins. Adding new features as often as I can.